Allow for widgets that don't have settings
My attempt at a way to allow widgets that don't have any settings to show up.

The change in 5.9 to isset within the display_callback made it so my settings array that has a value of null for the instance number hides my widget. I would like it if we allowed for objects and used isset for that but if the $instances is an array we can continue using array_key_exists so that widgets with no settings can work.